# users/openapi.yaml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
  description: API Gateway — handles sessions, auth, and proxies game manager traffic.
...
components:
  schemas:
    UserSession:
      type: object
      properties:
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          example: "johndoe@example.com"
    PlayCoordsResponse:
      type: object
      properties:
        coords:
          $ref: '#/components/schemas/Coordinates'
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
    Coordinates:
      type: object
      description: Barycentric coordinates on the triangular board (x + y + z == size - 1).
      required: [x, y, z]
      properties:
        x:
          type: integer
          example: 2
        y:
          type: integer
          example: 1
        z:
          type: integer
          example: 0
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessionId
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
paths:
...
  /api/me:
    get:
      tags: [Session]
      summary: Get the current authenticated user
      description: Returns the user data stored in the server-side Redis session.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Authenticated user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSession'
        '401':
          description: Not authenticated or session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Session store unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: >
        Creates a new account via the Auth service. On success a server-side Redis
        session is created and a `sessionId` cookie is set.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: "johndoe"
                email:
                  type: string
                  example: "johndoe@example.com"
                password:
                  type: string
                  example: "s3cr3tP@ss"
      responses:
        '200':
          description: User registered and session created
          headers:
            Set-Cookie:
              description: sessionId cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSession'
        '409':
          description: Email or username already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Auth service unreachable or session creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      description: >
        Authenticates the user via the Auth service. On success a server-side Redis
        session is created and a `sessionId` cookie is set.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: "johndoe@example.com"
                password:
                  type: string
                  example: "s3cr3tP@ss"
      responses:
        '200':
          description: Login successful, session created
          headers:
            Set-Cookie:
              description: sessionId cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSession'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Auth service unreachable or session creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/logout:
    post:
      tags: [Auth]
      summary: Log out the current user
      description: Deletes the Redis session and clears the `sessionId` cookie.
      security:
        - sessionCookie: []
        - csrfHeader: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out"
        '403':
          description: Invalid or missing CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/update-username:
    post:
      tags: [User]
      summary: Update the current user's username
      description: >
        Persists the new username in the Auth service (Firebase) and updates the
        Redis session so subsequent `/api/me` calls return the new value.
      security:
        - sessionCookie: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: "newusername"
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: "newusername"
        '400':
          description: Username field missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated or session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error while updating username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /play:
    get:
      tags: [ Interops ]
      summary: Ask the game engine for a bot move
      description: >
        Forwards the request to the game engine (`gamey`) and returns the coordinates
        the selected bot would play for the given YEN position.
        If `bot_id` is omitted the engine defaults to `minimax_bot`.
      parameters:
        - in: query
          name: position
          required: true
          description: URL-encoded JSON representing the current game state in YEN format.
          schema:
            type: string
          example: '{"size":3,"turn":0,"players":["B","R"],"layout":"./B./..."}'
        - in: query
          name: bot_id
          required: false
          description: Identifier of the bot to use. Available bots include `random_bot`, `greedy_bot` and `minimax_bot`.
          schema:
            type: string
            enum: [ random_bot, greedy_bot, minimax_bot ]
          example: "minimax_bot"
      responses:
        '200':
          description: Coordinates chosen by the bot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayCoordsResponse'
        '400':
          description: Invalid YEN position or malformed query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Game engine unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /createuser:
    post:
      tags: [Legacy]
      summary: Create a user (legacy endpoint)
      description: Legacy endpoint kept for backwards compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: "testuser"
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Hello testuser! welcome to the course!"
        '400':
          description: Error creating user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'